+++ __tests__/countdown.test.js
@@ -1,16 +1,15 @@
-import React from "react";
-import { render, screen, act } from "@testing-library/react";
-import Countdown from "../src/app/ui/Countdown";
-import "@testing-library/jest-dom";
+import { render, screen } from '@testing-library/react';
+import Countdown from '../src/app/ui/CountdownSection';
+import '@testing-library/jest-dom';
 
 jest.useFakeTimers();
 
-describe("Countdown component", () => {
+describe('Countdown component', () => {
   const mockData = {
-    targetDate: new Date("2025-06-19T00:00:00Z"),
+    targetDate: new Date('2025-06-19T00:00:00Z'),
   };
 
-  test("renders countdown with initial values", () => {
+  test('renders countdown with initial values', () => {
     render(<Countdown data={mockData} />);
 
     expect(screen.getByText(/Days/i)).toBeInTheDocument();
@@ -19,40 +18,38 @@ describe("Countdown component", () => {
     expect(screen.getByText(/Seconds/i)).toBeInTheDocument();
   });
 
-  test("should update countdown over time", () => {
-    // Mock the system time and set it to June 12th 2025 
-    jest.setSystemTime(new Date("2025-06-12T00:00:00Z"));
+  test('should update countdown over time', () => {
+    // Mock the system time and set it to June 12th 2025
+    jest.setSystemTime(new Date('2025-06-12T00:00:00Z'));
 
     // Render the Countdown component for the first time
     const { unmount } = render(<Countdown data={mockData} />);
-  
+
     // Retrieve the amount of days left on the countdown
-    const daysElement = screen.getByText("days").previousSibling;
+    const daysElement = screen.getByText('days').previousSibling;
     const initialDays = Number(daysElement.textContent);
 
     // Unmount the previously rendered Countdown component
     unmount();
-  
+
     // Update the mocked system time and set it to 6 days later
-    jest.setSystemTime(new Date("2025-06-18T00:00:00Z"));
+    jest.setSystemTime(new Date('2025-06-18T00:00:00Z'));
 
     // Render the Countdown component for the second time
     render(<Countdown data={mockData} />);
-  
+
     // Retrieve the amount of days left on the countdown with the new mocked system time in place
-    const updatedDaysElement = screen.getByText("days").previousSibling;
+    const updatedDaysElement = screen.getByText('days').previousSibling;
     const updatedDays = Number(updatedDaysElement.textContent);
-  
+
     // Expect the countdown to display fewer days than initially
     expect(updatedDays).toBeLessThan(initialDays);
   });
-  
-  
 
-  test("should not let Countdown component count below zero", () => {
+  test('should not let Countdown component count below zero', () => {
     render(<Countdown data={mockData} />);
 
-    jest.setSystemTime(new Date("2814-01-01T00:00:00Z"));
+    jest.setSystemTime(new Date('2814-01-01T00:00:00Z'));
 
     const timeLeftElements = screen.getAllByText(/\d+/);
 
@@ -62,15 +59,11 @@ describe("Countdown component", () => {
     });
   });
 
-  test("clears interval when unmounting", () => {
-    const clearIntervalMock = jest.spyOn(global, "clearInterval");
-
+  test('clears interval when unmounting', () => {
+    const clearIntervalMock = jest.spyOn(global, 'clearInterval');
     const { unmount } = render(<Countdown data={mockData} />);
-
     unmount();
-
     expect(clearIntervalMock).toHaveBeenCalled();
-
     clearIntervalMock.mockRestore();
   });
 });