+++ prettier.config.mjs
@@ -1,0 +1,13 @@
+/**
+ * @see https://prettier.io/docs/configuration
+ * @type {import("prettier").Config}
+ */
+const config = {
+  plugins: ['prettier-plugin-tailwindcss'],
+  semi: true,
+  singleQuote: true,
+  trailingComma: 'es5',
+  tabWidth: 2,
+};
+
+export default config;