+++ src/app/page.js
@@ -1,7 +1,8 @@
-import Countdown from "./ui/Countdown";
+import CountdownSection from "./ui/CountdownSection";
 import FaqSection from "./ui/FaqSection.jsx";
 import SignupSection from "./ui/SignupSection";
 import Navbar from "./ui/Navbar";
+import HeroSection from "./ui/HeroSection";
 import OrganiserSection from "./ui/OrganiserSection";
 import { getData } from "./service/FetchData";
 
@@ -12,14 +13,18 @@ export default async function Home() {
   const { navbar } = data;
 
   return (
-    <div>
-      <Navbar data={data.navbar} />
-      <main>
-        <FaqSection data={faqSection.faqs} />
-        <Countdown data={countdownData.countdown} />
-        <SignupSection data={signUpData.signUp} />
-        <OrganiserSection data={organisersSection.organisers} />
-      </main>
-    </div>
+    <>
+      <Navbar data={navbar} />
+      <div>
+        <Navbar data={data.navbar} />
+        <main>
+          <HeroSection data={heroData.hero} />
+          <FaqSection data={faqSection.faqs} />
+          <CountdownSection data={countdownData.countdown} />
+          <SignupSection data={signUpData.signUp} />
+          <OrganiserSection data={organisersSection.organisers} />
+        </main>
+      </div>
+    </>
   );
 }