+++ CHANGELOG.md
@@ -1,3 +1,23 @@
+## [0.1.20] - 2025-05-19
+
+### Added
+
+- Added custom Capathon colors to `styles.css`.
+
+## [0.1.19] - 2025-05-13
+
+### Fixed & added
+
+- Resolved code smells.
+- Added `prettier.config.mjs`.
+- Added prettier-plugin-tailwindcss plugin.
+
+## [0.1.18] - 2025-05-09
+
+### Added
+
+- Added `HeroSection.jsx` that renders titles and a highlight section.
+
 ## [0.1.17] - 2025-05-08
 
 ### Fixed